Temporary hack to PULSE model, MUST TRY TO FIX LATER
I've been banging my head on this for hours, trying to figure
out how to get eqc_temporal:stateful() to do what I want it to
do, but the damn thing is a mystery: the 2nd fun is being called
*much* more frequently than I believe it has any right to do.
stateful() is a black box, no source code, so figuring out WTF
it's working like it is ... maddening.

So, I've added a hack to ignore fold errors by forked pids.
BLEH.

It's certainly possible to know what the values were at the
very start of the fold/fold_keys.  The trick is to propagate
that info forward to the checking place.  {sigh}  That's a
fight for another day.