Remove use of tombstone2
The new non-backwards incompatible tombstone had information about the
file/offset of the value it was replacing so that a partial merge could
tell whether it should merge it or drop it.  It would drop it if the
value it was replacing was also merged out of existence. But this is not
safe. It only takes another level of indirection to break the logic. If
an even older value survives the merge that removes the tombstone and
the value the tombstone was deleting, re-opening could still resurrect
that value.  It is not possible to safely drop a tombstone in a partial
merge, so we're not doing that anymore. That may end up in extra disk
use that only a complete merge could reclaim, but it is the only safe
way to do it.

The out of date logic used during merged was tweaked a little to surface
what it was really doing: sometimes it flags a value as not as current
as the value in the keydir, sometimes it flagged a value that had
expired or a value not found on the keydir (which means a later
tombstone is present).