WIP Allow partial merge tombstone deletions
This introduces a new tombstone with a pointer to the file and offset of
the value being deleted. Currently only the file part is used. When
putting a value that replaces another value in a previous file, we now
write a tombstone for the previous value, then the new value.  When
merging and finding a tombstone, if the original file is part of the
merge, the tombstone can be dropped safely.  If not, then we re-open the
original file and append the tombstone to it, so that a future merge
hitting that file will drop it.

This passes all unit tests, but not pulse. Something is funky somewhere.