Drop tombstones on prefix merge and explicit merge
When merging a consecutive set of files starting from the first file in
the bitcask directory, it is safe to drop all encountered tombstones.
Detect this kind of "prefix" merge.
Allow merge input files to be specified by creating a merge.txt file in
the cask directory. bitcask:needs_merge() picks the contents of this
file to merge and will delete merge.txt when all files in it have been
deleted or marked for deletion.